+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # Changelog
 
+## Unreleased
+
+* refactor: allow to change max request body JSON nesting depth via injection
+
 ## 0.24.35 - 2024-01-09
 
 * fix: increase JSON request body max depth to allow for non-primitive attributes (i.e. array structures)